move volunteers components together
