+++ app/views/profile/ProfileView.js
@@ -3,8 +3,8 @@ import './divBox.css'
 import {Link} from 'react-router';
 import EventHistory from '../../components/EventDashboard/EventHistory'
 import UpcomingEvents from '../../components/EventDashboard/UpcomingEvents'
-import AboutMe from '../../components/aboutMe/AboutMe';
-import ProfileChart from '../../components/profileChart/ProfileChart';
+import AboutMe from '../../components/Volunteer/AboutMe';
+import ProfileChart from '../../components/Volunteer/ProfileChart';
 import FeaturedEvents from '../../containers/FeaturedEventsContainer';
 
 class ProfileView extends Component {
